Update Wizard messages && make page 5 consistent with the rest
