+++ app/lib/screens/wizard/page4.dart
@@ -11,9 +11,9 @@ class Page4 extends StatelessWidget {
       subtitle: 'NEWS',
       imagePath: 'assets/news_outline.png',
       description:
-          "Stay updated with ThreeFold's latest updates via the News section within the app.",
-      heightPercentage: 0.4,
-      widthPercentage: 0.75,
+          "Stay up to date with ThreeFolds' latest news",
+      heightPercentage: 0.35,
+      widthPercentage: 0.70,
     );
   }
 }
